refactor: merge `grid.spec` and `gridHelper.spec`
